Added and completed documentation in SDK.
